Fix malformed path by reducing path overlap
If source code in directory src/a, e.g. src/a/foo.c, is built using
a Makefile in src, then we may get a libtool dir of src/a/.libs and
a source file with the path a/foo.c. This cannot be joined
as src/a/a/foo.c but needs to be src/a/foo.c.

This patch implements a function that reduces path overlap between
two paths, such as the above libtool path and the source file path
so they can be joined correctly.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>