Run the regresion tests over multiple gcc versions
Whilst investigating issues with dodgey coverage being reported to
coveralls, it was found that the issue was dependent on the version of
g++/gcov used. Extending the existing test suite to multiple gcc
versions shows this up.

The issue is that in version 8, gcov changed the format of its output,
which makes the assumptions made in the current parser invalid. Tools
such as LCOV have already migrated to use gcov's new JSON format to work
around this issue.

For now the affected tests are marked as known failure to prevent a
failing build.