Send service_name and service_job_id only if both are set
By default we set `service_name` to `travis-ci` if `service_name` is
unset, but that in combination with an empty `service_job_id` causes all
such submitted results to end up in a single build (as they have the same
job ID) instead of the expected behavior where each submitted result has
a separate build.

The API[0] states that we should either use repo token or _both_ service
job ID and service name, so let's do just that.

[0] https://docs.coveralls.io/api-introduction